Refactor cookie utils
