+++ src/components/molecules/MockDataNotification.vue
@@ -1,6 +1,7 @@
 <script setup lang="ts">
 import { ref, onMounted } from 'vue'
 import { USE_MOCK_DATA } from '@/services'
+import { Cookie } from '@/utils/cookie'
 import CloseButton from '@/components/atoms/CloseButton.vue'
 
 const COOKIE_NAME = 'pmr_mock_notification_dismissed'
@@ -8,32 +9,13 @@ const COOKIE_DAYS = 7
 
 const isVisible = ref(true)
 
-const getCookie = (name: string): string | null => {
-  const nameEQ = name + '='
-  const cookies = document.cookie.split(';')
-  for (let cookie of cookies) {
-    cookie = cookie.trim()
-    if (cookie.indexOf(nameEQ) === 0) {
-      return cookie.substring(nameEQ.length)
-    }
-  }
-  return null
-}
-
-const setCookie = (name: string, value: string, days: number) => {
-  const date = new Date()
-  date.setTime(date.getTime() + days * 24 * 60 * 60 * 1000)
-  const expires = 'expires=' + date.toUTCString()
-  document.cookie = `${name}=${value};${expires};path=/`
-}
-
 const handleClose = () => {
   isVisible.value = false
-  setCookie(COOKIE_NAME, 'true', COOKIE_DAYS)
+  Cookie.set(COOKIE_NAME, 'true', COOKIE_DAYS)
 }
 
 onMounted(() => {
-  if (getCookie(COOKIE_NAME) === 'true') {
+  if (Cookie.get(COOKIE_NAME) === 'true') {
     isVisible.value = false
   }
 })
