+++ src/utils/cookie.ts
@@ -1,0 +1,19 @@
+export const Cookie = {
+  get: (name: string): string | null => {
+    const nameEQ = name + '='
+    const cookies = document.cookie.split(';')
+    for (let cookie of cookies) {
+      cookie = cookie.trim()
+      if (cookie.indexOf(nameEQ) === 0) {
+        return cookie.substring(nameEQ.length)
+      }
+    }
+    return null
+  },
+  set: (name: string, value: string, days: number) => {
+    const date = new Date()
+    date.setTime(date.getTime() + days * 24 * 60 * 60 * 1000)
+    const expires = 'expires=' + date.toUTCString()
+    document.cookie = `${name}=${value};${expires};path=/`
+  }
+}
